Fix warnings on Windows
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@689072 13f79535-47bb-0310-9956-ffa450edef68

+++ src/main/activemq/connector/openwire/OpenWireConnector.cpp
@@ -1027,7 +1027,7 @@ void OpenWireConnector::acknowledge( const SessionInfo* session,
             consumerInfo->getConsumerInfo()->getConsumerId()->cloneDataStructure() );
         ack.setDestination( amqMessage->getDestination()->cloneDataStructure() );
         ack.setLastMessageId( amqMessage->getMessageId()->cloneDataStructure() );
-        ack.setMessageCount( messages.size() );
+        ack.setMessageCount( (int)messages.size() );
 
         if( session->getAckMode() == cms::Session::SESSION_TRANSACTED ) {
